backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pikepdf==9.4.2
Pillow==11.0.0
pydantic==2.9.2
PyMuPDF==1.24.14
